rename : 유효성검사 함수파일 util로 이동
